fix(components): address review — wording, serialized geometry, ushort groups
- Tone down docs (README, CHANGELOG, sample README, api-coverage): stop
  claiming "full shape catalog" / "entire demo catalog"; frame the sample as a
  practical coverage check, not proof of complete Rapier API coverage.
- Make inline collider geometry persist in Scene/Prefab: add [SerializeField]
  to RapierConvexHullCollider.points, RapierTrimeshCollider.vertices/indices,
  and RapierVoxelsCollider.points.
- Avoid silent uint->ushort truncation: store collision/solver group
  memberships and filters as ushort on RapierColliderComponent.

+++ Packages/com.afjk.rapier/Runtime/Components/RapierColliderComponent.cs
@@ -20,11 +20,11 @@ namespace AFJK.Rapier
         [SerializeField] private RapierCoefficientCombineRule frictionCombineRule = RapierCoefficientCombineRule.Average;
         [SerializeField] private RapierCoefficientCombineRule restitutionCombineRule = RapierCoefficientCombineRule.Average;
         [SerializeField] private bool overrideCollisionGroups;
-        [SerializeField] private uint collisionGroupMemberships = 0xFFFF;
-        [SerializeField] private uint collisionGroupFilter = 0xFFFF;
+        [SerializeField] private ushort collisionGroupMemberships = 0xFFFF;
+        [SerializeField] private ushort collisionGroupFilter = 0xFFFF;
         [SerializeField] private bool overrideSolverGroups;
-        [SerializeField] private uint solverGroupMemberships = 0xFFFF;
-        [SerializeField] private uint solverGroupFilter = 0xFFFF;
+        [SerializeField] private ushort solverGroupMemberships = 0xFFFF;
+        [SerializeField] private ushort solverGroupFilter = 0xFFFF;
         [SerializeField] private RapierActiveEvents activeEvents = RapierActiveEvents.None;
         [SerializeField] private bool overrideActiveCollisionTypes;
         [SerializeField] private RapierActiveCollisionTypes activeCollisionTypes =
@@ -245,14 +245,14 @@ namespace AFJK.Rapier
             {
                 activeWorld.SetColliderCollisionGroups(
                     ColliderHandle,
-                    RapierWorld.InteractionGroups((ushort)collisionGroupMemberships, (ushort)collisionGroupFilter));
+                    RapierWorld.InteractionGroups(collisionGroupMemberships, collisionGroupFilter));
             }
 
             if (overrideSolverGroups)
             {
                 activeWorld.SetColliderSolverGroups(
                     ColliderHandle,
-                    RapierWorld.InteractionGroups((ushort)solverGroupMemberships, (ushort)solverGroupFilter));
+                    RapierWorld.InteractionGroups(solverGroupMemberships, solverGroupFilter));
             }
 
             if (activeEvents != RapierActiveEvents.None)
